Improved guests' commodities view, so they can see their booking even when it hasn't started yet
Instead of showing them their booking only when it has started, now they can see the nearest one in time, as long as it hasn´t been already finished.

Example:
 - It's 12th March and the guest has two bookings ahead:
   - From 10th April to 17th April
   - From 05th June to 08th June
They'll now see the information related to the first booking, and once it's finished, the information will switch to the second one.